requirements.txt

torch==2.3.1+cu118 
torchaudio==2.3.1+cu118 
torchvision==0.18.1+cu118
numpy>=1.16.5
scipy>=1.3.0
opencv-python
gdown
pillow
gradio
timm==0.4.12
wandb
matplotlib
